+++ src/transformers/index.ts
@@ -17,4 +17,5 @@ export * as popover from './popover.js';
 export * as tabs from './tabs.js';
 export * as tag from './tag.js';
 export * as tooltip from './tooltip.js';
+export * as tree from './tree.js';
 export * as video from './video.js';